requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
Babel==2.13.1
beautifulsoup4==4.12.2
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorlog==6.8.0
confection==0.1.4
cssselect==1.2.0
cymem==2.0.8
distro==1.8.0
exceptiongroup==1.2.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2023.12.0
gast==0.4.0
google-auth==2.24.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==7.0.0
jax==0.4.20
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.2
keras==2.12.0
langcodes==3.3.0
langdetect==1.0.9
libclang==16.0.6
lxml==4.9.3
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.3.1
murmurhash==1.0.10
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cudnn-cu11==8.6.0.163
oauthlib==3.2.2
openai==1.3.7
opt-einsum==3.3.0
packaging==23.2
Pillow==10.1.0
preshed==3.0.9
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
thinc==8.2.1
threadpoolctl==3.2.0
tinysegmenter==0.3
tldextract==5.1.1
tokenizers==0.15.0
tqdm==4.66.1
transformers==4.35.2
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
